refactor: atualizando a estrutura do projeto
