+++ src/app.module.ts
@@ -1,18 +1,15 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { ClienteModule } from './cliente/cliente.module';
-import { ClienteController } from './cliente/cliente.controller';
-import { ClienteService } from './cliente/cliente.service';
-import { GerenteController } from './gerente/gerente.controller';
-import { GerenteService } from './gerente/gerente.service';
-import { GerenteModule } from './gerente/gerente.module';
-import { ContaModule } from './conta/conta.module';
-
+import { ClienteModule } from './modules/cliente.module';
+import { ClienteController } from './controllers/cliente.controller';
+import { ClienteService } from './services/cliente.service';
+import { GerenteController } from './controllers/gerente.controller';
+import { GerenteService } from './services/gerente.service';
+import { GerenteModule } from './modules/gerente.module';
+import { ContaModule } from './modules/conta.module';
 
 @Module({
   imports: [ClienteModule, GerenteModule, ContaModule],
-  controllers: [AppController, ClienteController, GerenteController],
-  providers: [AppService, ClienteService, GerenteService],
+  controllers: [ ClienteController, GerenteController],
+  providers: [ ClienteService, GerenteService],
 })
 export class AppModule {}
