+++ src/models/cliente.model.ts
@@ -1,5 +1,5 @@
-import { Conta } from "src/conta/conta.model";
-import { InterfacePessoa } from "src/interfaces/pessoa.interface";
+import { Conta } from '../models/contas/conta.model';
+import { InterfacePessoa } from '../interfaces/pessoa.interface';
 
 export class Cliente implements InterfacePessoa {
   constructor(
@@ -15,5 +15,5 @@ export class Cliente implements InterfacePessoa {
     public rendaSalarial: number,
     public conta: Conta[] = []
   ){}
-
+  
 }
